# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API Playlist Vidoe",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/scholarshipinformation": {
            "get": {
                "description": "list scholarshipinformation entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List scholarshipinformation entities",
                "operationId": "list-scholarshipinformation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Scholarshipinformation"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create scholarshipinformation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create scholarshipinformation",
                "operationId": "create-scholarshipinformation",
                "parameters": [
                    {
                        "description": "Scholarshipinformation entity",
                        "name": "playlist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Scholarshipinformation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Scholarshipinformation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/scholarshipinformation/{id}": {
            "get": {
                "description": "get scholarshipinformation by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a scholarshipinformation entity by ID",
                "operationId": "get-scholarshipinformation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Scholarshipinformation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Scholarshipinformation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/scholarshiprequests": {
            "get": {
                "description": "list scholarshiprequest entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List scholarshiprequest entities",
                "operationId": "list-scholarshiprequest",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.ScholarshipRequest"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create scholarshiprequest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create scholarshiprequest",
                "operationId": "create-scholarshiprequest",
                "parameters": [
                    {
                        "description": "ScholarshipRequest entity",
                        "name": "scholarshiprequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Scholarshiprequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.ScholarshipRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.Scholarshiprequest": {
            "type": "object",
            "properties": {
                "scholarshipinformation": {
                    "type": "integer"
                },
                "scholarshiptype": {
                    "type": "integer"
                },
                "semester": {
                    "type": "integer"
                },
                "user": {
                    "type": "integer"
                }
            }
        },
        "ent.ScholarshipRequest": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ScholarshipRequestQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ScholarshipRequestEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ScholarshipRequestEdges": {
            "type": "object",
            "properties": {
                "scholarshipinformation": {
                    "description": "Scholarshipinformation holds the value of the Scholarshipinformation edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Scholarshipinformation"
                },
                "scholarshiptype": {
                    "description": "Scholarshiptype holds the value of the Scholarshiptype edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Scholarshiptype"
                },
                "semester": {
                    "description": "Semester holds the value of the Semester edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Semester"
                },
                "user": {
                    "description": "User holds the value of the User edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
        "ent.Scholarshipinformation": {
            "type": "object",
            "properties": {
                "Scholarship_name": {
                    "description": "ScholarshipName holds the value of the \"Scholarship_name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ScholarshipinformationQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ScholarshipinformationEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ScholarshipinformationEdges": {
            "type": "object",
            "properties": {
                "scholarshipRequest": {
                    "description": "ScholarshipRequest holds the value of the ScholarshipRequest edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.ScholarshipRequest"
                    }
                },
                "scholarshiptype": {
                    "description": "Scholarshiptype holds the value of the Scholarshiptype edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Scholarshiptype"
                }
            }
        },
        "ent.Scholarshiptype": {
            "type": "object",
            "properties": {
                "Type_name": {
                    "description": "TypeName holds the value of the \"Type_name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ScholarshiptypeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ScholarshiptypeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ScholarshiptypeEdges": {
            "type": "object",
            "properties": {
                "scholarshipRequest": {
                    "description": "ScholarshipRequest holds the value of the ScholarshipRequest edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.ScholarshipRequest"
                    }
                },
                "scholarshipinformation": {
                    "description": "Scholarshipinformation holds the value of the Scholarshipinformation edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Scholarshipinformation"
                    }
                }
            }
        },
        "ent.Semester": {
            "type": "object",
            "properties": {
                "Semester": {
                    "description": "Semester holds the value of the \"Semester\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SemesterQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SemesterEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.SemesterEdges": {
            "type": "object",
            "properties": {
                "scholarshipRequest": {
                    "description": "ScholarshipRequest holds the value of the ScholarshipRequest edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.ScholarshipRequest"
                    }
                }
            }
        },
        "ent.User": {
            "type": "object",
            "properties": {
                "Email": {
                    "description": "Email holds the value of the \"Email\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.UserEdges": {
            "type": "object",
            "properties": {
                "scholarshipRequest": {
                    "description": "ScholarshipRequest holds the value of the ScholarshipRequest edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.ScholarshipRequest"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}